refactor: Change database -> period_data
